gameref convenience trait methods terrible migration

+++ src/probability.rs
@@ -21,22 +21,22 @@ impl <T> RandomEventGenerator<T> where T: EventSpawner {
     }
 }
 
-pub fn individual_event(probability: f64) -> bool {
-    random::<f64>() < probability
+pub fn individual_event(probability: f32) -> bool {
+    random::<f32>() < probability
 }
 
-pub fn logistic(x: f64) -> f64 {
+pub fn logistic(x: f32) -> f32 {
     0.5 + 0.5 * (x / 2.0).tanh()
 }
 
-pub fn dev_mean_sample(stddev: f64, mean: f64) -> f64 {
-    thread_rng().sample::<f64, StandardNormal>(StandardNormal) * stddev + mean
+pub fn dev_mean_sample(stddev: f32, mean: f32) -> f32 {
+    thread_rng().sample::<f32, StandardNormal>(StandardNormal) * stddev + mean
 }
 
 pub fn positive_isample(stddev: usize, mean: usize) -> usize {
-    dev_mean_sample(stddev as f64, mean as f64).max(0.0).round() as usize
+    dev_mean_sample(stddev as f32, mean as f32).max(0.0).round() as usize
 }
 
-pub fn sample(stddev: f64) -> f64 {
+pub fn sample(stddev: f32) -> f32 {
     dev_mean_sample(stddev, 0.0)
 }
